model-endpoint/requirements.txt
runpod==1.2.3
torch==2.2.0
Pillow==10.0.0
numpy==1.24.0
torchvision==0.17.0
